Refactor class UnixSocket method remove().

+++ src/unixsocket.cpp
@@ -15,7 +15,7 @@
 
 #include "network/unixsocket.h"                 // UnixSocket
 #include "network/handle-type.h"                // handle_type
-#include "network/pathname.h"                   // Pathname
+#include "network/optionalpathname.h"           // Pathname
 #include "network/socket-family-type.h"         // socket_family_type
 #include "network/to-path.h"                    // to_path()
 
@@ -47,16 +47,20 @@ auto Network::UnixSocket::open(const ByteString& t_addr,
     return result;
 }
 
-auto Network::UnixSocket::remove(const Pathname& pathname) const -> void
+auto Network::UnixSocket::remove(const OptionalPathname& pathname) const -> void
 {
+    if (!pathname) {
+        return;
+    }
+
     if (is_verbose()) {
         std::cout << "Calling std::filesystem::remove("
-                  << pathname
+                  << *pathname
                   << ')'
                   << std::endl;
     }
 
-    std::filesystem::remove(pathname);
+    std::filesystem::remove(*pathname);
 }
 
 auto Network::UnixSocket::state(SocketState t_state) -> UnixSocket&
@@ -69,9 +73,7 @@ auto Network::UnixSocket::state(SocketState t_state) -> UnixSocket&
     }
 
     if (previous == SocketState::bound) {
-        if (const auto pathname {to_path(sockname())}) {
-            remove(*pathname);
-        }
+        remove(to_path(sockname()));
     }
 
     m_state = next;
